backend/requirements.txt
## Backend runtime dependencies (used to create the .venv)
# Keep this file minimal — used by CI and local dev to install packages.
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
python-dotenv
